Updating docs and renaming charts by removing 'component' from the name
