+++ src/charts/index.js
@@ -1,5 +1,5 @@
-export {default as LegendComponent} from './legend/LegendComponent.js';
-export {default as TooltipComponent} from './tooltip/TooltipComponent.js';
-export {default as BarComponent} from './bar/BarComponent.js';
-export {default as DonutComponent} from './donut/DonutComponent.js';
-export {default as StackedAreaComponent} from './stackedArea/StackedAreaComponent.js';
+export {default as Legend} from './legend/Legend.js';
+export {default as Tooltip} from './tooltip/Tooltip.js';
+export {default as Bar} from './bar/Bar.js';
+export {default as Donut} from './donut/Donut.js';
+export {default as StackedArea} from './stackedArea/StackedArea.js';
