+++ src/charts/legend/Legend.test.js
@@ -2,20 +2,20 @@ import React from 'react';
 import Enzyme, { shallow, mount } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-15';
 
-import BarComponent from './BarComponent';
-import barData from './barChart.fixtures';
+import Legend from './Legend';
+import legendData from './legendChart.fixtures';
 
-import bar from './barChart';
+import legendChart from './legendChart';
 
 Enzyme.configure({ adapter: new Adapter() });
 
-describe('Bar Chart Component', () => {
+describe('Legend Chart', () => {
 
     describe('render', () => {
 
         describe('when data is not passed', () => {
             it('should throw an error', () => {
-                expect(() => shallow(<BarComponent data={{}} />)).toThrow();
+                expect(() => shallow(<Legend data={[]} />)).toThrow();
             });
         });
 
@@ -23,7 +23,7 @@ describe('Bar Chart Component', () => {
             let createSpy;
 
             beforeEach(() => {
-                createSpy = jest.spyOn(bar, 'create');
+                createSpy = jest.spyOn(legendChart, 'create');
             });
 
             afterEach(() => {
@@ -32,64 +32,64 @@ describe('Bar Chart Component', () => {
             });
 
             it('should call the create method or the chart', () => {
-                mount(<BarComponent chart={bar} data={barData.withLetters()} />);
+                mount(<Legend chart={legendChart} data={legendData.with6Points()} />);
 
-                const expected = 1;
-                const actual = createSpy.mock.calls.length;
+                let expected = 1;
+                let actual = createSpy.mock.calls.length;
 
                 expect(actual).toEqual(expected);
             });
 
             it('should call the create method or the chart with the container as the first argument', () => {
-                const wrapper = mount(<BarComponent chart={bar} data={barData.withLetters()} />);
+                const wrapper = mount(<Legend chart={legendChart} data={legendData.with6Points()} />);
 
-                const expected = wrapper.find('.bar-container').instance();
-                const actual = createSpy.mock.calls[0][0];
+                let expected = wrapper.find('.legend-container').instance();
+                let actual = createSpy.mock.calls[0][0];
 
                 expect(actual).toEqual(expected);
             });
 
             it('should call the create method or the chart with the configuration object as the second argument', () => {
-                const dataSet = barData.withLetters();
+                const dataSet = legendData.with6Points();
 
-                mount(<BarComponent chart={bar} data={dataSet} />);
+                mount(<Legend chart={legendChart} data={dataSet} />);
 
-                const expectedData = dataSet;
-                const actualData = createSpy.mock.calls[0][1];
+                let expectedData = dataSet;
+                let actualData = createSpy.mock.calls[0][1];
 
                 expect(actualData).toEqual(expectedData);
             });
 
             it('should allow setting width', () => {
-                const dataSet = barData.withLetters();
-                const expected = 500;
+                const dataSet = legendData.with6Points();
+                let expected = 500;
 
                 mount(
-                    <BarComponent
-                        chart={bar}
+                    <Legend
+                        chart={legendChart}
                         data={dataSet}
                         width={expected}
                     />
                 );
 
-                const actual = createSpy.mock.calls[0][2].width;
+                let actual = createSpy.mock.calls[0][2].width;
 
                 expect(actual).toEqual(expected);
             });
 
             it('should allow setting height', () => {
-                const dataSet = barData.withLetters();
-                const expected = 500;
+                const dataSet = legendData.with6Points();
+                let expected = 500;
 
                 mount(
-                    <BarComponent
-                        chart={bar}
+                    <Legend
+                        chart={legendChart}
                         data={dataSet}
                         height={expected}
                     />
                 );
 
-                const actual = createSpy.mock.calls[0][2].height;
+                let actual = createSpy.mock.calls[0][2].height;
 
                 expect(actual).toEqual(expected);
             });
@@ -102,7 +102,7 @@ describe('Bar Chart Component', () => {
             let updateSpy;
 
             beforeEach(() => {
-                updateSpy = jest.spyOn(bar, 'update');
+                updateSpy = jest.spyOn(legendChart, 'update');
             });
 
             afterEach(() => {
@@ -111,35 +111,35 @@ describe('Bar Chart Component', () => {
             });
 
             it('should call the update method or the chart', () => {
-                const wrapper = mount(<BarComponent chart={bar} data={barData.withLetters()} />);
+                const wrapper = mount(<Legend chart={legendChart} data={legendData.with6Points()} />);
 
                 // Changing properties should trigger a componentDidUpdate
                 wrapper.setProps({
-                    data: barData.withColors(),
+                    data: legendData.with6Points(),
                 });
 
-                const expected = 1;
-                const actual = updateSpy.mock.calls.length;
+                let expected = 1;
+                let actual = updateSpy.mock.calls.length;
 
                 expect(actual).toEqual(expected);
             });
 
             it('should pass in the new data to the update method', () => {
-                const wrapper = mount(<BarComponent chart={bar} data={barData.withLetters()} />);
+                const wrapper = mount(<Legend chart={legendChart} data={legendData.with6Points()} />);
 
                 // Changing properties should trigger a componentDidUpdate
                 wrapper.setProps({
-                    data: barData.withColors(),
+                    data: legendData.with6Points(),
                 });
 
-                const expected = barData.withColors().length;
-                const actual = updateSpy.mock.calls[0][1].length;
+                let expected = legendData.with6Points().length;
+                let actual = updateSpy.mock.calls[0][1].length;
 
                 expect(actual).toEqual(expected);
             });
 
             it('should pass in the new configuration to the update method', () => {
-                const wrapper = mount(<BarComponent chart={bar} data={barData.withLetters()} />);
+                const wrapper = mount(<Legend chart={legendChart} data={legendData.with6Points()} />);
                 const expected = 20;
 
                 // Changing properties should trigger a componentDidUpdate
@@ -147,7 +147,7 @@ describe('Bar Chart Component', () => {
                     width: expected,
                 });
 
-                const actual = updateSpy.mock.calls[0][2].width;
+                let actual = updateSpy.mock.calls[0][2].width;
 
                 expect(actual).toEqual(expected);
             });
@@ -155,24 +155,24 @@ describe('Bar Chart Component', () => {
     });
 
     describe('unmount', () => {
-        let createSpy;
+        let destroySpy;
 
         beforeEach(() => {
-            createSpy = jest.spyOn(bar, 'destroy');
+            destroySpy = jest.spyOn(legendChart, 'destroy');
         });
 
         afterEach(() => {
-            createSpy.mockReset();
-            createSpy.mockRestore();
+            destroySpy.mockReset();
+            destroySpy.mockRestore();
         });
 
         it('should call the destroy method or the chart', () => {
-            const wrapper = mount(<BarComponent chart={bar} data={barData.withLetters()} />);
+            const wrapper = mount(<Legend chart={legendChart} data={legendData.with6Points()} />);
 
             wrapper.unmount();
 
-            const expected = 1;
-            const actual = createSpy.mock.calls.length;
+            let expected = 1;
+            let actual = destroySpy.mock.calls.length;
 
             expect(actual).toEqual(expected);
         });
